requirements.txt
aiogram==3.10.0
aiohttp==3.9.1
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.1.0
beautifulsoup4==4.12.2
redis==5.0.1
gigachat==0.1.11